// go-grpc-angular/src/app/grpc/SittelleConst_pb.d.ts
export enum EnumStatsAggregate { 
  STATS_AGGREAGATE_UNDEFINED = 0,
  STATS_AGGREAGATE_LEN = 15,
  STATS_AGGREAGATE_COUNT = 1,
  STATS_AGGREAGATE_DATE_CHANGE_MIN = 2,
  STATS_AGGREAGATE_DATE_CHANGE_MAX = 3,
  STATS_AGGREAGATE_DATE_PLANNING_MIN = 4,
  STATS_AGGREAGATE_DATE_PLANNING_MAX = 5,
  STATS_AGGREAGATE_PRIORITY_MIN = 6,
  STATS_AGGREAGATE_PRIORITY_MAX = 7,
  STATS_AGGREAGATE_TIME_ESTIMATE_MIN = 8,
  STATS_AGGREAGATE_TIME_ESTIMATE_MAX = 9,
  STATS_AGGREAGATE_TIME_ESTIMATE_COUNT = 10,
  STATS_AGGREAGATE_PRIORITY_STATUS_MIN = 11,
  STATS_AGGREAGATE_PRIORITY_STATUS_MAX = 12,
  STATS_AGGREAGATE_DATE_START_MIN = 13,
  STATS_AGGREAGATE_DATE_START_MAX = 14,
}
// ...
export enum EnumTaskAction { 
  TASK_ACTION_CREATE = 0,
  TASK_ACTION_STATUS_CHANGE = 1,
  TASK_ACTION_MANUAL_CHANGE = 2,
  TASK_ACTION_ROLE_CHANGE = 3,
  TASK_ACTION_CONTENT_CHANGE = 4,
  TASK_ACTION_COMMENT_ADD = 5,
  TASK_ACTION_STATE_CHANGE = 6,
  TASK_ACTION_ITEM_CHANGE = 7,
  TASK_ACTION_PARENT_CHANGE = 8,
  TASK_ACTION_STATUS_INPUT_CHANGE = 9,
  TASK_ACTION_LOST_ACCESS = 10,
  TASK_ACTION_ACCOUNT_STATE_CHANGE = 11,
}